+++ TechArtSurvey/Server/RedTeam.TechArtSurvey.WebApi/App_Start/Startup.cs
@@ -5,15 +5,15 @@ using JetBrains.Annotations;
 using Microsoft.Owin;
 using Microsoft.Owin.Security.OAuth;
 using Owin;
+using RedTeam.TechArtSurvey.WebApi.App_Start;
 using RedTeam.TechArtSurvey.WebApi.Authorization;
-using RedTeam.TechArtSurvey.WebApi.Conf;
 using RedTeam.TechArtSurvey.WebApi.DI;
 using RedTeam.TechArtSurvey.WebApi.Owin;
 
 [assembly: OwinStartup(typeof(Startup))]
 [assembly: log4net.Config.XmlConfigurator(Watch = true)]
 
-namespace RedTeam.TechArtSurvey.WebApi.Conf
+namespace RedTeam.TechArtSurvey.WebApi.App_Start
 {
     public class Startup
     {
